requirements.txt
seldonian-engine>=0.8.10
torch==2.1.0
transformers==4.34.0